put sampler boilerplate in separate files
